Implement `clocks_get_res` and `clocks_get_time`
This implements `clocks_get_res and `clocks_get_time` supporting
realtime and monotonic clocks with millisecond and microsecond
resolution respectively.

Process and thread clocks are implemented as aliases for the monotonic
clock because it is based on `performance.now` which counts from the
start of a process or thread process (worker) so it is the only time
origin available to count from.