Move exports into an instantiate-able module class
Originally I had intended for the main module to be compatible with
esm-integration imports but it turns out import-maps have been put
behind an unstable flag so they cannot be the main delivery method.

This puts all exports on an exports property of a `Module` class which
matches the interface provided by `WebAssembl.Instance` and will allow
each WebAssembly module to programmatically get their own instance of
our module with relative ease.

Support for import-maps and esm-integration can be done in a separate
module with relative ease as-well, it would just have to instantiate a
single instance of the module.