Make `env` a required property of `Module`
The `env` property on instances of `Module` is currently optional,
mostly just due to how the old tests were written which makes it
somewhat tedious to deal with.

This initializes it to an empty object if no option is provided and
removes the null checks that was present in the exports.