Add support for removing directories
This implements `path_remove_directory` by mapping it to `Deno.remove`.

For the sake of conformance, the path is checked to ensure it is
actually a directory before `Deno.remove` is called which introduces
some extra overhead in the name of conformance that we may not want to
keep in the long run.